qrcode_scanner/requirements.txt
opencv-python==4.7.0.72
pyzbar==0.1.9
requests==2.28.2
Pillow==9.5.0